Refactor/React Context API for all state management and functions
