+++ src/App.js
@@ -1,7 +1,7 @@
 // TODO:
-// local storage for user settings
 // incorporate styled components
 // more react-y things to enhance performance and readability
+// // Context API for state management
 
 import React from 'react';
 import Pomodoro from './Pomodoro';
