+++ src/Pomodoro/Settings/GoalSetter.js
@@ -1,32 +1,19 @@
 import React, { Component } from 'react';
+import TimersContext from '../TimersContext';
 
 // settings component - set a pomodoro goal
-class GoalSetter extends Component {
-  constructor(props) {
-    super(props);
-    this.handleGoalChange = this.handleGoalChange.bind(this);
-  }
-
-  handleGoalChange(change) {
-    const goal = this.props.goal + change;
-    if (goal < 1) return;
-    this.props.changeState({ goal });
-
-    // continue recursing the function every 0.1 seconds if mouse click is held
-    setTimeout(() => {
-      if (this.props.mouseDown) this.handleGoalChange(change);
-    }, 100);
-  }
-
+export default class GoalSetter extends Component {
   render() {
     return (
-      <div className='settings-item settings-goal'>
-          <a className='decrement noselect' onMouseDown={() => this.handleGoalChange(-1)}>–</a>
-          <div className='settings-goal-show'>Goal : {this.props.goal}</div>
-          <a className='increment noselect' onMouseDown={() => this.handleGoalChange(+1)}>+</a>
-        </div>
+      <TimersContext.Consumer>
+        {context => (
+          <div className='settings-item settings-goal'>
+              <a className='decrement noselect' onMouseDown={() => context.handleGoalChange(-1)}>–</a>
+              <div className='settings-goal-show'>Goal : {context.state.goal}</div>
+              <a className='increment noselect' onMouseDown={() => context.handleGoalChange(+1)}>+</a>
+            </div>
+        )}
+      </TimersContext.Consumer>
     )
   }
 }
-
-export default GoalSetter;
