+++ src/Pomodoro/ShowTime.js
@@ -1,10 +1,12 @@
 import React, { Component } from 'react';
+import TimersContext from './TimersContext';
 import './css/showTime.css';
 
 // displays the timer's current time formatted
 class ShowTime extends Component {
   render() {
-    const timeRemaining = this.props.timeRemaining;
+    const timeRemaining = 
+      this.props.context.state.activeTimer.timeRemaining;
 
     let mins = Math.floor((timeRemaining/1000)/60);
     if (mins < 10) mins = `0${mins}`; // prepend 0 if < 10
@@ -31,4 +33,10 @@ class ShowTime extends Component {
   )}
 }
 
-export default ShowTime;
+const WithContext = () => (
+  <TimersContext.Consumer>
+    {context => <ShowTime context={context} />}
+  </TimersContext.Consumer>
+)
+
+export default WithContext;
