+++ src/index.css
@@ -1,8 +1,3 @@
-/* Responsive screen sizes for media queries:
-420px
-768
-920 */
-
   /* --------------------------------------------------------------------------- */
  /*                                GENERAL STYLES                               */
 /* --------------------------------------------------------------------------- */
