+++ src/media-queries.css
@@ -1,4 +1,3 @@
-
   /* --------------------------------------------------------------------------- */
  /*                                MEDIA QUERIES                                */
 /* --------------------------------------------------------------------------- */
